Check 4
All code is done, finished all methods of the OOP, and added the main loop which prints menu, asks for user inputs, checks if user input is valid, and then calls the corresponding method. Also fixed bugs in the previous code. Next step is to add all comments explaining code.